New API refactor (#1283)
getenv -> getEnvironment
stoi -> from_chars

+++ src/frontend/ibusfrontend/ibusfrontend.cpp
@@ -11,6 +11,7 @@
 #include <sys/wait.h>
 #include <unistd.h>
 #include <cerrno>
+#include <charconv>
 #include <csignal>
 #include <cstddef>
 #include <cstdint>
@@ -20,6 +21,8 @@
 #include <optional>
 #include <set>
 #include <string>
+#include <string_view>
+#include <system_error>
 #include <tuple>
 #include <utility>
 #include <vector>
@@ -32,7 +35,9 @@
 #include "fcitx-utils/dbus/objectvtable.h"
 #include "fcitx-utils/dbus/servicewatcher.h"
 #include "fcitx-utils/dbus/variant.h"
+#include "fcitx-utils/environ.h"
 #include "fcitx-utils/event.h"
+#include "fcitx-utils/eventloopinterface.h"
 #include "fcitx-utils/flags.h"
 #include "fcitx-utils/handlertable.h"
 #include "fcitx-utils/key.h"
@@ -54,9 +59,8 @@
 #include "fcitx/misc_p.h"
 #include "dbus_public.h"
 
-FCITX_DEFINE_LOG_CATEGORY(ibus, "ibus")
-#define FCITX_IBUS_DEBUG() FCITX_LOGC(::ibus, Debug)
-#define FCITX_IBUS_WARN() FCITX_LOGC(::ibus, Warn)
+#define FCITX_IBUS_DEBUG() FCITX_LOGC(::fcitx::ibus, Debug)
+#define FCITX_IBUS_WARN() FCITX_LOGC(::fcitx::ibus, Warn)
 
 #define IBUS_INPUTMETHOD_DBUS_INTERFACE "org.freedesktop.IBus"
 #define IBUS_INPUTCONTEXT_DBUS_INTERFACE "org.freedesktop.IBus.InputContext"
@@ -70,34 +74,30 @@ namespace fcitx {
 
 namespace {
 
+FCITX_DEFINE_LOG_CATEGORY(ibus, "ibus")
+
 std::string getSocketPath(bool isWayland) {
-    auto *path = getenv("IBUS_ADDRESS_FILE");
+    auto path = getEnvironment("IBUS_ADDRESS_FILE");
     if (path) {
-        return path;
+        return *path;
     }
     std::string hostname = "unix";
     std::string displaynumber = "0";
     if (isWayland) {
         displaynumber = "wayland-0";
-        if (auto *display = getenv("WAYLAND_DISPLAY")) {
-            displaynumber = display;
+        if (auto display = getEnvironment("WAYLAND_DISPLAY")) {
+            displaynumber = *display;
         }
-    } else if (auto *display = getenv("DISPLAY")) {
-        auto *p = display;
-        for (; *p != ':' && *p != '\0'; p++) {
-            ;
-        }
+    } else if (auto displayEnv = getEnvironment("DISPLAY")) {
+        std::string_view display = *displayEnv;
+        auto colon = display.find(':');
 
-        char *displaynumberStart = nullptr;
-        if (*p == ':') {
-            hostname = std::string(display, p);
-            displaynumberStart = p + 1;
+        if (colon != std::string_view::npos) {
+            hostname = display.substr(0, colon);
+            auto displaynumberStart = display.substr(colon + 1);
 
-            for (; *p != '.' && *p != '\0'; p++) {
-                ;
-            }
-
-            displaynumber = std::string(displaynumberStart, p);
+            auto period = displaynumberStart.find('.');
+            displaynumber = displaynumberStart.substr(0, period);
         } else {
             hostname = display;
         }
@@ -124,9 +124,9 @@ std::pair<std::string, pid_t> getAddress(const std::string &socketPath) {
     pid_t pid = -1;
 
     /* get address from env variable */
-    auto *address = getenv("IBUS_ADDRESS");
+    auto address = getEnvironment("IBUS_ADDRESS");
     if (address) {
-        return {address, -1};
+        return {*address, -1};
     }
 
     /* read address from ~/.config/ibus/bus/soketfile */
@@ -140,14 +140,14 @@ std::pair<std::string, pid_t> getAddress(const std::string &socketPath) {
     const auto *pidValue = config.valueByPath("IBUS_DAEMON_PID");
 
     if (value && pidValue) {
-        try {
-            pid = std::stoi(*pidValue);
+        if (std::from_chars(pidValue->c_str(),
+                            pidValue->c_str() + pidValue->size(), pid)
+                .ec == std::errc()) {
             // Check if we are in flatpak, or pid is same with ourselves, or
             // another running process.
             if (isInFlatpak() || pid == getpid() || kill(pid, 0) == 0) {
                 return {*value, pid};
             }
-        } catch (...) {
         }
     }
 
@@ -176,12 +176,118 @@ pid_t runIBusExit() {
 
     return child_pid;
 }
-} // namespace
+
+std::optional<std::pair<std::string, pid_t>>
+readIBusInfo(const std::string &socketPath) {
+    std::pair<std::string, pid_t> address = getAddress(socketPath);
+
+    FCITX_IBUS_DEBUG() << "Found ibus address from file " << socketPath << ": "
+                       << address;
+
+    if (isInFlatpak()) {
+        // Check the in flatpak special pid value.
+        if (address.second == 0) {
+            return std::nullopt;
+        }
+    } else {
+        // It's not meaningful to compare pid from different pid namespace.
+        if (address.second == getpid()) {
+            return std::nullopt;
+        }
+    }
+    if (address.first.empty() ||
+        address.first.find("fcitx_random_string") != std::string::npos) {
+        return std::nullopt;
+    }
+    return address;
+}
+
+std::optional<std::pair<std::string, pid_t>>
+readIBusInfo(const std::set<std::string> &socketPaths) {
+    for (const auto &path : socketPaths) {
+        auto address = readIBusInfo(path);
+        if (address) {
+            return address;
+        }
+    }
+    return std::nullopt;
+}
+
+std::set<std::string> allSocketPaths(const StandardPath &standardPath) {
+    std::set<std::string> paths;
+    if (isInFlatpak()) {
+        // Flatpak always use DISPLAY=:99, which means we will need to guess
+        // what files are available.
+        auto map = standardPath.locateWithFilter(
+            StandardPath::Type::Config, "ibus/bus",
+            [](const std::string &path, const std::string &, bool user) {
+                if (!user) {
+                    return false;
+                }
+                return stringutils::startsWith(path, getLocalMachineId());
+            });
+
+        for (const auto &item : map) {
+            paths.insert(item.second);
+        }
+
+        // Make the guess that display is 0, it is the most common value that
+        // people would have.
+        if (paths.empty()) {
+            auto configHome =
+                standardPath.userDirectory(StandardPath::Type::Config);
+            if (!configHome.empty()) {
+                paths.insert(stringutils::joinPath(
+                    configHome, "ibus/bus",
+                    stringutils::concat(getLocalMachineId(), "-unix-", 0)));
+            }
+        }
+    } else {
+        if (auto path = getFullSocketPath(standardPath, false); !path.empty()) {
+            paths.insert(std::move(path));
+        }
+    }
+
+    // Also add wayland.
+    if (auto path = getFullSocketPath(standardPath, true); !path.empty()) {
+        paths.insert(std::move(path));
+    }
+    return paths;
+}
+
+constexpr uint32_t releaseMask = (1 << 30);
 
 using AttachmentsType = FCITX_STRING_TO_DBUS_TYPE("a{sv}");
 using IBusText = FCITX_STRING_TO_DBUS_TYPE("(sa{sv}sv)");
 using IBusAttrList = FCITX_STRING_TO_DBUS_TYPE("(sa{sv}av)");
 using IBusAttribute = FCITX_STRING_TO_DBUS_TYPE("(sa{sv}uuuu)");
+
+IBusAttribute makeIBusAttr(uint32_t type, uint32_t value, uint32_t start,
+                           uint32_t end) {
+    IBusAttribute attr;
+    std::get<0>(attr) = "IBusAttribute";
+    std::get<2>(attr) = type;
+    std::get<3>(attr) = value;
+    std::get<4>(attr) = start;
+    std::get<5>(attr) = end;
+    return attr;
+}
+
+IBusAttrList makeIBusAttrList() {
+    IBusAttrList attrList;
+    std::get<0>(attrList) = "IBusAttrList";
+    return attrList;
+}
+
+IBusText makeSimpleIBusText(const std::string &str) {
+    IBusText text;
+    std::get<0>(text) = "IBusText";
+    std::get<2>(text) = str;
+    std::get<3>(text).setData(makeIBusAttrList());
+    return text;
+}
+
+} // namespace
 
 class IBusFrontend : public dbus::ObjectVTable<IBusFrontend> {
 public:
@@ -210,33 +316,6 @@ private:
     dbus::Bus *bus_;
     std::unique_ptr<dbus::ServiceWatcher> watcher_;
 };
-
-constexpr uint32_t releaseMask = (1 << 30);
-
-IBusAttribute makeIBusAttr(uint32_t type, uint32_t value, uint32_t start,
-                           uint32_t end) {
-    IBusAttribute attr;
-    std::get<0>(attr) = "IBusAttribute";
-    std::get<2>(attr) = type;
-    std::get<3>(attr) = value;
-    std::get<4>(attr) = start;
-    std::get<5>(attr) = end;
-    return attr;
-}
-
-IBusAttrList makeIBusAttrList() {
-    IBusAttrList attrList;
-    std::get<0>(attrList) = "IBusAttrList";
-    return attrList;
-}
-
-IBusText makeSimpleIBusText(const std::string &str) {
-    IBusText text;
-    std::get<0>(text) = "IBusText";
-    std::get<2>(text) = str;
-    std::get<3>(text).setData(makeIBusAttrList());
-    return text;
-}
 
 class IBusInputContext;
 class IBusService : public dbus::ObjectVTable<IBusService> {
@@ -651,48 +730,6 @@ IBusFrontend::createInputContext(const std::string & /* unused */) {
     return ic->path();
 }
 
-std::set<std::string> allSocketPaths(const StandardPath &standardPath) {
-    std::set<std::string> paths;
-    if (isInFlatpak()) {
-        // Flatpak always use DISPLAY=:99, which means we will need to guess
-        // what files are available.
-        auto map = standardPath.locateWithFilter(
-            StandardPath::Type::Config, "ibus/bus",
-            [](const std::string &path, const std::string &, bool user) {
-                if (!user) {
-                    return false;
-                }
-                return stringutils::startsWith(path, getLocalMachineId());
-            });
-
-        for (const auto &item : map) {
-            paths.insert(item.second);
-        }
-
-        // Make the guess that display is 0, it is the most common value that
-        // people would have.
-        if (paths.empty()) {
-            auto configHome =
-                standardPath.userDirectory(StandardPath::Type::Config);
-            if (!configHome.empty()) {
-                paths.insert(stringutils::joinPath(
-                    configHome, "ibus/bus",
-                    stringutils::concat(getLocalMachineId(), "-unix-", 0)));
-            }
-        }
-    } else {
-        if (auto path = getFullSocketPath(standardPath, false); !path.empty()) {
-            paths.insert(std::move(path));
-        }
-    }
-
-    // Also add wayland.
-    if (auto path = getFullSocketPath(standardPath, true); !path.empty()) {
-        paths.insert(std::move(path));
-    }
-    return paths;
-}
-
 IBusFrontendModule::IBusFrontendModule(Instance *instance)
     : instance_(instance), socketPaths_(allSocketPaths(standardPath_)) {
     dbus::VariantTypeRegistry::defaultRegistry().registerType<IBusText>();
@@ -763,42 +800,6 @@ IBusFrontendModule::~IBusFrontendModule() {
 
 dbus::Bus *IBusFrontendModule::bus() {
     return dbus()->call<IDBusModule::bus>();
-}
-
-std::optional<std::pair<std::string, pid_t>>
-readIBusInfo(const std::string &socketPath) {
-    std::pair<std::string, pid_t> address = getAddress(socketPath);
-
-    FCITX_IBUS_DEBUG() << "Found ibus address from file " << socketPath << ": "
-                       << address;
-
-    if (isInFlatpak()) {
-        // Check the in flatpak special pid value.
-        if (address.second == 0) {
-            return std::nullopt;
-        }
-    } else {
-        // It's not meaningful to compare pid from different pid namespace.
-        if (address.second == getpid()) {
-            return std::nullopt;
-        }
-    }
-    if (address.first.empty() ||
-        address.first.find("fcitx_random_string") != std::string::npos) {
-        return std::nullopt;
-    }
-    return address;
-}
-
-std::optional<std::pair<std::string, pid_t>>
-readIBusInfo(const std::set<std::string> &socketPaths) {
-    for (const auto &path : socketPaths) {
-        auto address = readIBusInfo(path);
-        if (address) {
-            return address;
-        }
-    }
-    return std::nullopt;
 }
 
 void IBusFrontendModule::replaceIBus(bool recheck) {
@@ -964,7 +965,7 @@ void IBusFrontendModule::becomeIBus(bool recheck) {
             return true;
         });
 
-    auto that = this;
+    auto *that = this;
     that->timeEvent_ = std::move(timeEvent);
 }
